Added Inject_DLL and str_to_wchar_t functions
Inject_DLL: responsible for injecting a DLL into the target process. [in] hProcess - Handle to target process; [in] DLL_Path - Path to DLL to be injected; [out] ThreadID - Pointer to Thread Identifier. str_to_wchar_t: takes an std::string and returns a pointer to a wide string. [in] src - source string to convert to *wchar_t